Move LSLStreamRecorder to an importable Python file
